various fixes and tweaks to mons and early rewards

+++ include/multiboot_pokemon_colosseum.h
@@ -1,6 +1,6 @@
 #ifndef GUARD_MULTIBOOT_POKEMON_COLOSSEUM_H
 #define GUARD_MULTIBOOT_POKEMON_COLOSSEUM_H
 
-extern const u16 gMultiBootProgram_PokemonColosseum_Start[0x14000];
+extern const u32 gMultiBootProgram_PokemonColosseum_Start[0x14000];
 
 #endif // GUARD_MULTIBOOT_POKEMON_COLOSSEUM_H
